refactor(api): projects members 라우트를 전용 모듈로 추출
5개 라우트를 api/projects/members.py로 이동. 핸들러 본문·경로·함수명 무변경 —
sed 로 원문 구간을 그대로 잘라 옮겼다. 라우트 테이블 characterization 통과
(14건 동일, 가림 쌍 0건).

_get_admin_org_ids 를 _shared.py 로 승격했다. registry 2회 + members 1회에
더해 패키지 밖 세 모듈(project_configs·claude_sessions·routes)이 쓰는 유일한
공유 이름이다. 함수 본문이 전부 함수 내부 import 라 _shared.py 는 모듈 레벨
의존이 0이며, 형제 모듈을 import 하지 않는다(의존은 항상 형제 → _shared).

도메인 경계가 데이터 모델 경계와 일치한다 — members 는 ProjectMember*·
OrgMember* 6종만, registry 는 DBProject* 만 쓴다. 공유는 위 헬퍼 하나뿐이다.

grep 심볼 카운트가 sqlalchemy.delete 를 과다 집계했고 ruff 가 잡았다.
분할 시 import 목록의 진실원은 grep 이 아니라 린터다.

게이트: mypy 0, pytest 1344 passed(알려진 로컬 플레이크 1건 제외).
